解决所有参数传递问题，已正常输出

+++ fortran_prj/Cbellhoplib/test/inc/readin.h
@@ -25,10 +25,10 @@ struct DigitalPara{
     float rloop;
     float deptht;
     struct Complex cpt;
-    float rhot;
+    double rhot;
     float depthb;
     struct Complex cpb;
-    float rhob;
+    double rhob;
 };
